% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{as.Continuous}
\alias{as.Continuous}
\title{Redefine a random variable as a Continuous r.v.}
\usage{
as.Continuous(x, npoints = 9, logtol = -8)
}
\arguments{
\item{x}{}

\item{npoints}{}

\item{logtol}{}
}
\description{
Redefine a random variable as a Continuous r.v.
}

