% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\docType{class}
\name{Constant-class}
\alias{Constant}
\alias{Constant-class}
\title{Constant}
\description{
Constant
}
\section{Methods}{

\describe{
\item{\code{cdf(x = NULL, lower.tail = TRUE, log.p = FALSE, ...)}}{Computes cdf(x)}

\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE, conf = NULL)}}{Default method to initialize basic objects}

\item{\code{invcdf(p = NULL, lower.tail = TRUE, log.p = FALSE)}}{Computes quantiles associated with input probability(ies)}

\item{\code{pdf(x = NULL, log = FALSE, ...)}}{Computes pdf(x)}

\item{\code{rand(n = 1)}}{Generates n replicates of this object's value}
}}

