% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{\%=\%}
\alias{\%=\%}
\title{Temporarily fixing a random variable at a constant (e.g., an observation),
or a Constant at a different value}
\usage{
l \%=\% r
}
\arguments{
\item{l, }{a random variable or Constant (class rcvirtual.random)}

\item{r, }{a numeric, matrix, or Constant}
}
\value{
x, a Constant
}
\description{
Temporarily fixing a random variable at a constant (e.g., an observation),
or a Constant at a different value
}

