% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\docType{class}
\name{Normal-class}
\alias{Normal}
\alias{Normal-class}
\title{Normal random variable}
\description{
Normal random variable
}
\section{Methods}{

\describe{
\item{\code{cdf(x = NULL, lower.tail = TRUE, log.p = FALSE, verbose = FALSE, ...)}}{Computes the Normal cdf(x)}

\item{\code{given(variable, value, permanent = FALSE)}}{If permanent = FALSE, returns a new variable conditioned on a fixed value
of one of the its dependencies. If permanent = TRUE, modifies the current
variable (i.e., .self) permanently.}

\item{\code{iget.optimization.parameters()}}{This function should be replaced with another
that does not return NULL, when auxiliary
parameters exist.}

\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE, conf = NULL)}}{Default method to initialize basic objects}

\item{\code{invcdf(p = NULL, lower.tail = TRUE, log.p = FALSE, verbose = FALSE)}}{Computes quantiles associated with input probability(ies)}

\item{\code{pdf(x = NULL, log = FALSE, verbose = FALSE, ...)}}{Provides either the pdf or its evaluation for given x}

\item{\code{rand(n = 1, from.marginal = FALSE)}}{Generates n (multivariate) normal variates}
}}

