% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{+.rcvirtual.random}
\alias{+.rcvirtual.random}
\title{Allowed operations with random abstract types}
\usage{
\method{+}{rcvirtual.random}(l, r)
}
\arguments{
\item{l}{}

\item{r}{}
}
\description{
Allowed operations with random abstract types
}

